move css into folders and create a components folder
