+++ client/src/App.js
@@ -1,23 +1,12 @@
 import React from 'react';
-import logo from './logo.svg';
-import './App.css';
+
+import './css/App.css';
 
 function App() {
   return (
     <div className="App">
       <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
+        <h1>Recommendation App</h1>
       </header>
     </div>
   );
